requirements.txt
pydantic
pydantic_ai
spotipy
python-dotenv
duckdb